Renamed 'filehandler' package to 'io' and created new classes in it.
