+++ src/main/java/io/DataLoader.java
@@ -1,4 +1,4 @@
-package filehandler;
+package io;
 
 import java.io.BufferedInputStream;
 import java.io.BufferedReader;
