+++ src/main/java/io/FileInputFormat.java
@@ -1,0 +1,12 @@
+package io;
+
+import java.nio.file.Path;
+
+import adoop.Job;
+
+public class FileInputFormat {
+
+	public static void addInputPath(Job job, Path inputPath) {
+		job.setInputPath(inputPath);
+	}
+}
